Move dir. /libs/reactQuery.ts -> /libs/reactQuery/
